# api.yml
openapi: 3.0.0

info:
  title: Trekking Management API
  version: 1.0.0
  description: API documentation for Trekking Management Application

paths:

  /:
    get:
      summary: Home page

  /register:
    get:
      summary: Render user registration page
    post:
      summary: Register a new user

  /staff_register:
    get:
      summary: Render trek staff registration page
    post:
      summary: Register a new trek staff

  /login:
    get:
      summary: Render login page
    post:
      summary: Authenticate user or trek staff

  /admin_login:
    get:
      summary: Render admin login page
    post:
      summary: Authenticate admin

  /logout:
    get:
      summary: Logout current user

  /user_dashboard:
    get:
      summary: View user dashboard

  /user/apply/{trek_id}:
    post:
      summary: Book a trek
      parameters:
        - name: trek_id
          in: path
          required: true
          schema:
            type: integer

  /user/edit_profile/{id}:
    get:
      summary: Render edit profile page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
    post:
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /user/bookings:
    get:
      summary: View user booking history

  /staff_dashboard:
    get:
      summary: View trek staff dashboard

  /staff/my_treks:
    get:
      summary: View assigned treks

  /staff/trek/{trek_id}:
    get:
      summary: View trek details
      parameters:
        - name: trek_id
          in: path
          required: true
          schema:
            type: integer

  /staff/remove_participant/{booking_id}:
    post:
      summary: Remove participant from trek
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer

  /staff/edit_staff:
    get:
      summary: Render edit staff profile page
    post:
      summary: Update staff profile

  /staff/trek/{id}/booking_status:
    post:
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /staff/trek/{id}/slots:
    get:
      summary: Render update slots page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
    post:
      summary: Update available slots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /staff/trek/{id}/status:
    post:
      summary: Update trek status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin_dashboard:
    get:
      summary: View admin dashboard

  /admin/users:
    get:
      summary: View all users

  /admin/users/blacklist/{id}:
    get:
      summary: Blacklist user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/users/approve/{id}:
    get:
      summary: Approve user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/staff:
    get:
      summary: View all trek staff

  /admin/staff/added/{id}:
    get:
      summary: Approve trek staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/staff/remove/{id}:
    get:
      summary: Remove trek staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/staff/approve/{id}:
    get:
      summary: Remove staff from blacklist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/staff/blacklist/{id}:
    get:
      summary: Blacklist trek staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/treks:
    get:
      summary: View all treks

  /admin/trek/create_trek:
    get:
      summary: Render create trek page
    post:
      summary: Create a new trek

  /admin/trek/edit/{id}:
    get:
      summary: Render edit trek page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
    post:
      summary: Update trek details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/trek/remove/{id}:
    get:
      summary: Remove trek
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer

  /admin/bookings:
    get:
      summary: View all bookings